% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{aes}
\alias{aes}
\title{Define aesthetic mappings}
\usage{
aes(...)
}
\arguments{
\item{...}{Named arguments mapping aesthetics to variables}
}
\value{
An aesthetic mapping object
}
\description{
Define aesthetic mappings
}
\examples{
aes(x = mpg, y = hp)
aes(x = mpg, fill = cyl)
}
