% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R
\name{api}
\alias{api}
\title{Programmatic API for rMosaic}
\description{
This file provides a programmatic R interface for creating Mosaic visualizations.
Users can use R functions instead of writing JSON/YAML specs.
}
\keyword{internal}
