% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{bin}
\alias{bin}
\title{Bin a continuous variable}
\usage{
bin(field, ...)

bin(field, ...)
}
\arguments{
\item{column}{Column name to bin}

\item{step}{Optional bin step size}
}
\value{
A binning channel
}
\description{
Bin a continuous variable

Bin a field
}
