% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{brush_x}
\alias{brush_x}
\title{Add X-axis brush to plot}
\usage{
brush_x(selection)
}
\arguments{
\item{selection}{A selection created with brush_selection()}
}
\description{
Add X-axis brush to plot
}
