% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{brush_xy}
\alias{brush_xy}
\title{Add 2D brush to plot}
\usage{
brush_xy(selection)
}
\arguments{
\item{selection}{A selection created with brush_selection()}
}
\description{
Add 2D brush to plot
}
