% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{brush_y}
\alias{brush_y}
\title{Add Y-axis brush to plot}
\usage{
brush_y(selection)
}
\arguments{
\item{selection}{A selection created with brush_selection()}
}
\description{
Add Y-axis brush to plot
}
