% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{build_spec}
\alias{build_spec}
\title{Convert ggmosaic to Mosaic spec}
\usage{
build_spec(p)
}
\arguments{
\item{p}{A ggmosaic object}
}
\value{
A list with spec and data components
}
\description{
Convert ggmosaic to Mosaic spec
}
