% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{color_domain_fixed}
\alias{color_domain_fixed}
\title{Fix color domain (prevents auto-scaling across facets)}
\usage{
color_domain_fixed()
}
\description{
Fix color domain (prevents auto-scaling across facets)
}
