% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{color_scheme}
\alias{color_scheme}
\title{Set color scheme}
\usage{
color_scheme(scheme)
}
\arguments{
\item{scheme}{Color scheme (e.g., "blues", "oranges", "viridis")}
}
\description{
Set color scheme
}
