% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{count}
\alias{count}
\title{Count aggregation}
\usage{
count(field = "")

count(field = "")
}
\value{
A count aggregation channel
}
\description{
Count aggregation

Count aggregation
}
