% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{data_table}
\alias{data_table}
\title{Data table bound to a selection}
\usage{
data_table(from, columns, as, filter = NULL, width = 680, height = 215)
}
\arguments{
\item{from}{Table name}

\item{columns}{Columns to display}

\item{as}{Selection the table writes to}

\item{filter}{Selection to filter by}

\item{width}{Table width}

\item{height}{Table height}
}
\description{
Data table bound to a selection
}
