% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{dot}
\alias{dot}
\title{Create a dot mark (scatterplot)}
\usage{
dot(data, channels, ...)

dot(data, channels, ...)
}
\arguments{
\item{data}{Data reference from from()}

\item{encoding}{Named list of channel encodings}
}
\value{
A mark specification
}
\description{
Create a dot mark (scatterplot)

Dot mark
}
