% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\docType{data}
\name{Fixed}
\alias{Fixed}
\title{Fixed domain constant}
\format{
An object of class \code{mz_constant} of length 1.

An object of class \code{mz_constant} of length 1.
}
\usage{
Fixed

Fixed
}
\description{
Fixed domain constant

Constant for fixed domains (equivalent to JSON value "Fixed")
}
\keyword{datasets}
