% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{fixed_domain}
\alias{fixed_domain}
\title{Set fixed domain}
\usage{
fixed_domain(axis = c("x", "y", "both"))
}
\arguments{
\item{axis}{Which axis ("x", "y", or "both")}
}
\description{
Set fixed domain
}
