% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{from}
\alias{from}
\title{Reference a data source}
\usage{
from(table, filterBy = NULL)

from(table, filterBy = NULL)
}
\arguments{
\item{table}{Table name in the data list}

\item{filterBy}{Optional selection to filter by}
}
\value{
A data reference object
}
\description{
Reference a data source

Reference a data source with optional filter
}
\examples{
\dontrun{
from("mtcars")
brush <- Selection$crossfilter()
from("mtcars", filterBy = brush)
}
}
