% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_area}
\alias{geom_area}
\title{Area layer}
\usage{
geom_area(
  mapping = aes(),
  data = NULL,
  fill = "steelblue",
  alpha = 0.5,
  filter = NULL,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Override default data}

\item{fill}{Fill color}

\item{alpha}{Opacity}

\item{filter}{Selection to filter by}

\item{...}{Additional parameters}
}
\description{
Area layer
}
