% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_bar}
\alias{geom_bar}
\title{Bar chart layer}
\usage{
geom_bar(mapping = aes(), data = NULL, fill = "steelblue", filter = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Override default data}

\item{fill}{Fill color}

\item{filter}{Selection to filter by}

\item{...}{Additional parameters}
}
\description{
Bar chart layer
}
