% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_density2d}
\alias{geom_density2d}
\title{Heatmap/density layer}
\usage{
geom_density2d(mapping = aes(), data = NULL, filter = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Override default data}

\item{filter}{Selection to filter by}

\item{...}{Additional parameters}
}
\description{
Heatmap/density layer
}
