% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_histogram}
\alias{geom_histogram}
\title{Histogram layer}
\usage{
geom_histogram(
  mapping = aes(),
  data = NULL,
  bins = 30,
  binwidth = NULL,
  fill = "steelblue",
  filter = NULL,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings (aes)}

\item{data}{Override default data}

\item{bins}{Number of bins (default: 30)}

\item{binwidth}{Bin width}

\item{fill}{Fill color}

\item{filter}{Selection to filter by}

\item{...}{Additional mark parameters}
}
\description{
Histogram layer
}
\examples{
\dontrun{
ggmosaic(mtcars, aes(x = mpg)) +
  geom_histogram(fill = "steelblue")
}
}
