% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_histogram_y}
\alias{geom_histogram_y}
\title{Vertical histogram (counts on x, bins on y)}
\usage{
geom_histogram_y(
  mapping = aes(),
  data = NULL,
  bins = 30,
  binwidth = NULL,
  fill = "steelblue",
  filter = NULL,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings (aes)}

\item{data}{Override default data}

\item{bins}{Number of bins (default: 30)}

\item{binwidth}{Bin width}

\item{fill}{Fill color}

\item{filter}{Selection to filter by}

\item{...}{Additional mark parameters}
}
\description{
Vertical histogram (counts on x, bins on y)
}
