% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_line}
\alias{geom_line}
\title{Line layer}
\usage{
geom_line(
  mapping = aes(),
  data = NULL,
  color = "steelblue",
  size = 1.5,
  filter = NULL,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Override default data}

\item{color}{Line color}

\item{size}{Line width}

\item{filter}{Selection to filter by}

\item{...}{Additional parameters}
}
\description{
Line layer
}
