% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_point}
\alias{geom_point}
\title{Point/dot layer}
\usage{
geom_point(
  mapping = aes(),
  data = NULL,
  size = 3,
  alpha = 0.7,
  filter = NULL,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Override default data}

\item{size}{Point size (default: 3)}

\item{alpha}{Opacity (0-1)}

\item{filter}{Selection to filter by}

\item{...}{Additional parameters}
}
\description{
Point/dot layer
}
