% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{geom_raster}
\alias{geom_raster}
\title{Tile/raster layer}
\usage{
geom_raster(mapping = aes(), data = NULL, filter = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Override default data}

\item{filter}{Selection to filter by}

\item{...}{Additional parameters}
}
\description{
Tile/raster layer
}
