% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R,
%   R/ggmosaic_enhanced.R
\name{hconcat}
\alias{hconcat}
\title{Horizontal concatenation of plots}
\usage{
hconcat(...)

hconcat(...)

hconcat(...)
}
\arguments{
\item{...}{ggmosaic plots or layout objects to concatenate horizontally}
}
\value{
A layout specification

A layout object
}
\description{
Horizontal concatenation of plots

Horizontal concatenation

Horizontal concatenation of plots
}
\examples{
\dontrun{
p1 <- ggmosaic(mtcars, aes(x = mpg)) + geom_histogram()
p2 <- ggmosaic(mtcars, aes(x = hp)) + geom_histogram()
hconcat(p1, p2) \%>\% render()
}
}
