% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{hspace}
\alias{hspace}
\title{Add horizontal spacing}
\usage{
hspace(pixels = 10)
}
\arguments{
\item{pixels}{Number of pixels for horizontal spacing}
}
\value{
A spacing element
}
\description{
Add horizontal spacing
}
