% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index.R
\name{Index}
\alias{Index}
\title{Returns the mosaic database}
\usage{
Index(i)
}
\arguments{
\item{i}{Version of the MOSAIC database}
}
\value{
mosaic database
}
\description{
This downloads the MOSAIC database as an S4 object
}
\examples{
mos_fetch("v1.0.0")
}
