% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{input_menu}
\alias{input_menu}
\title{Menu input bound to a selection}
\usage{
input_menu(from, column, label, as)
}
\arguments{
\item{from}{Table name}

\item{column}{Column to drive the menu}

\item{label}{Label text}

\item{as}{Selection created with brush_selection()}
}
\description{
Menu input bound to a selection
}
