% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{intervalX}
\alias{intervalX}
\title{Create an X-axis interval brush}
\usage{
intervalX(as)

intervalX(as)
}
\arguments{
\item{as}{Selection to bind to}
}
\value{
An interactor specification
}
\description{
Create an X-axis interval brush

Interval selection along X
}
