% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{intervalXY}
\alias{intervalXY}
\title{Create a 2D interval brush}
\usage{
intervalXY(as)

intervalXY(as)
}
\arguments{
\item{as}{Selection to bind to}
}
\value{
An interactor specification
}
\description{
Create a 2D interval brush

2D interval selection
}
