% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{intervalY}
\alias{intervalY}
\title{Create a Y-axis interval brush}
\usage{
intervalY(as)

intervalY(as)
}
\arguments{
\item{as}{Selection to bind to}
}
\value{
An interactor specification
}
\description{
Create a Y-axis interval brush

Interval selection along Y
}
