% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{legend_color}
\alias{legend_color}
\title{Color legend for a plot}
\usage{
legend_color(for_plot, columns = 1, as = NULL)
}
\arguments{
\item{for_plot}{Name of the plot to attach legend to}

\item{columns}{Number of legend columns}

\item{as}{Optional selection to filter by}
}
\description{
Color legend for a plot
}
