% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R
\name{line}
\alias{line}
\title{Create a line mark}
\usage{
line(data, encoding = list())
}
\arguments{
\item{data}{Data reference from from()}

\item{encoding}{Named list of channel encodings}
}
\value{
A mark specification
}
\description{
Create a line mark
}
