% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{list_mosaic_selections}
\alias{list_mosaic_selections}
\title{List all stored Mosaic selections}
\usage{
list_mosaic_selections()
}
\value{
Character vector of selection names.
}
\description{
Returns the names of all selections currently held in the package store.
}
\seealso{
\code{\link{get_mosaic_selection}}
}
