% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R
\name{mean_agg}
\alias{mean_agg}
\title{Mean aggregation}
\usage{
mean_agg(column)
}
\arguments{
\item{column}{Column to average}
}
\value{
A mean aggregation channel
}
\description{
Mean aggregation
}
