% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_metadata.R
\name{mos_fetch}
\alias{mos_fetch}
\title{Returns mosaic database}
\usage{
mos_fetch(version)
}
\arguments{
\item{version}{version of mosaic database. currently "v1.0.0".}
}
\value{
mosaic database
}
\description{
This function returns the mosaic database
}
\examples{
mos_fetch("v1.0.0")
}
