% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_metadata.R
\name{mos_metadata}
\alias{mos_metadata}
\title{Returns metadata for a given trait and species}
\usage{
mos_metadata(trait, index)
}
\arguments{
\item{trait}{Trait name of a factorial variable. Search mos_traitAllSpp(mosaic) for trait list if needed.}

\item{index}{Index of the species of interest search "which(slot(mosaic, "species")==BinomialName)" for index.}
}
\value{
dataframe of metadata for a given species and trait record
}
\description{
This function returns a dataframe containing the metadata for a given record (trait + spp) - including
author, year, journal, database, and mosaic. See the MOSAIC user guide for more information.
See mos_multiMetaRecords if searching metadata for > 1 species.
}
\examples{
mos_metadata("volancy", 14)
}
