% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_multiMetaRecords.R
\name{mos_multiMetaRecords}
\alias{mos_multiMetaRecords}
\title{Returns metadata for a given trait and >1 species}
\usage{
mos_multiMetaRecords(trait, sequenceOfIds)
}
\arguments{
\item{trait}{Trait name of a factorial variable. Search mos_traitAllSpp(mosaic) for trait list if needed.}

\item{index}{Index of the species of interest search "match(slot(mosaic, "species"), c(BinomialName1,BinomialName1, etc.))" for indices}
}
\value{
dataframe of metadata for a given species and trait record
}
\description{
This function returns a dataframe containing the metadata for a given record (trait + multiple spp) - including
author, year, journal, database, and mosaic. See the MOSAIC user guide for more information.
}
\examples{
mos_multiMetaRecords("volancy", 14)
}
