% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_singleSppTraitSummary.R
\name{mos_singleSppTraitSummary}
\alias{mos_singleSppTraitSummary}
\title{Returns all MOSAIC attribute/trait records for one species}
\usage{
mos_singleSppTraitSummary(speciesNameLatin)
}
\arguments{
\item{speciesNameLatin}{Binomial species name (Genus species).}
}
\value{
list of records (no metadata) for all attribute/trait fields in MOSAIC for the species
}
\description{
This function returns all records (no metadata) in the MOSAIC database for one species.
}
\examples{
mos_singleSppTraitSummary("Alces alces")
}
