% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_sppAllTrait.R
\name{mos_sppAllTrait}
\alias{mos_sppAllTrait}
\title{Returns all MOSAIC attribute/trait records for one or more species}
\usage{
mos_sppAllTrait(sppList)
}
\arguments{
\item{sppList}{list of one or more species for which a trait summary is desired.}
}
\value{
list of records (no metadata) for all attribute/trait fields in MOSAIC
}
\description{
This function returns all records (no metadata) in the MOSAIC database for one or more species.
}
\examples{
mos_sppAllTrait("Alces alces")
mos_sppAllTrait(("Acinonyx jubatus", "Acropora downingi", "Aepyceros melampus", "Alces alces", "Alligator mississippiensis")))
}
