% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_traitAllSpp.R
\name{mos_traitAllSpp}
\alias{mos_traitAllSpp}
\title{Returns all records in MOSAIC for a given trait}
\usage{
mos_traitAllSpp(trait)
}
\arguments{
\item{trait}{Name of trait. Prompt:  mos_traitList(mosaic)  for the list of trait names.}
}
\value{

}
\description{
This function returns all records in the MOSAIC database for a given trait.
}
\examples{
mos_sppCheck("biomass")
mos_sppCheck("sexual dimorphism")
mos_sppCheck("volancy")
}
