% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_traitFrequency.R
\name{mos_traitFrequency}
\alias{mos_traitFrequency}
\title{Returns the absolute and relative occurance of any given trait level.}
\usage{
mos_traitFrequency(trait)
}
\arguments{
\item{trait}{Trait name of a factorial variable. Search mos_traitAllSpp(mosaic) for trait list if needed.}
}
\value{
list of records (no metadata) for all attribute/trait fields in MOSAIC for the species
}
\description{
This function returns the frequency and counts of records for factorial variable (e.g., 30\% volant, 70\% non-volant)
}
\examples{
mos_traitFrequency("volancy")
mos_traitFrequency("sexual dimorphism")
}
