% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_traitNames.R
\name{mos_traitNames}
\alias{mos_traitNames}
\title{Reports the trait names from the mosaic object}
\usage{
mos_traitNames(mosaic)
}
\arguments{
\item{mosaic}{The mosaic object (run:  mosaic <- mos_fetch("v1.0.0")  to obtain mosaic)}
}
\value{

}
\description{
This function extracts the traitnames from the mosaic object
}
\examples{
mos_traitNames(mosaic)
}
