% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos_fetch.R
\docType{class}
\name{mosaic_meta-class}
\alias{mosaic_meta-class}
\alias{mosaic_meta}
\title{Returns the mosaic database}
\arguments{
\item{versionSet}{Version of the MOSAIC database}
}
\value{
mosaic database
}
\description{
This downloads the MOSAIC database as an S4 object
}
\examples{
mos_fetch("v1.0.0")
}
