% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R
\name{new_mosaic_element}
\alias{new_mosaic_element}
\title{Create a new mosaic element}
\usage{
new_mosaic_element(spec, class = NULL)
}
\description{
Create a new mosaic element
}
\keyword{internal}
