% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{plot_name}
\alias{plot_name}
\title{Set a name on a plot for cross-references (e.g., legends)}
\usage{
plot_name(p, name)
}
\arguments{
\item{p}{ggmosaic plot}

\item{name}{plot name}
}
\description{
Set a name on a plot for cross-references (e.g., legends)
}
