% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{plot_size}
\alias{plot_size}
\title{Set plot size}
\usage{
plot_size(width = NULL, height = NULL)
}
\arguments{
\item{width}{Width in pixels}

\item{height}{Height in pixels}
}
\description{
Set plot size
}
