% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{+.ggmosaic}
\alias{+.ggmosaic}
\title{Add components to a ggmosaic plot}
\usage{
\method{+}{ggmosaic}(e1, e2)
}
\arguments{
\item{e1}{A ggmosaic object}

\item{e2}{A layer, scale, or other component}
}
\description{
Add components to a ggmosaic plot
}
