% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\name{rectY}
\alias{rectY}
\title{Create a rectY mark (bar chart/histogram)}
\usage{
rectY(data, channels, ...)

rectY(data, channels, ...)
}
\arguments{
\item{data}{Data reference from from()}

\item{encoding}{Named list of channel encodings (x, y, fill, etc.)}
}
\value{
A mark specification
}
\description{
Create a rectY mark (bar chart/histogram)

Rectangular bars along Y (histogram/bar)
}
\examples{
\dontrun{
rectY(
  from("mtcars"),
  list(x = bin("mpg"), y = count(), fill = "steelblue")
)
}
}
