% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderMosaic}
\alias{renderMosaic}
\title{Shiny render function for Mosaic}
\usage{
renderMosaic(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a call to \code{\link{mosaic}()}.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use this in the server to render the Mosaic visualization to the output.
}
