% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.except.R
\name{rm.except}
\alias{rm.except}
\title{Returns the mosaic database}
\usage{
rm.except(except, pattern)
}
\arguments{
\item{except}{Version of the MOSAIC database}

\item{pattern}{Version of the MOSAIC database}
}
\value{
mosaic database
}
\description{
This downloads the MOSAIC database as an S4 object
}
\examples{
mos_fetch("v1.0.0")
}
