% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R
\docType{data}
\name{Selection}
\alias{Selection}
\title{Selection class for interactive selections}
\format{
An object of class \code{list} of length 4.

An object of class \code{list} of length 4.
}
\usage{
Selection

Selection
}
\arguments{
\item{type}{Selection type (crossfilter, single, intersect, union)}

\item{name}{Optional name for the selection parameter}
}
\description{
Create interactive selections for filtering and linking plots.

Create a new Selection

Get spec representation

Helpers to declare Mosaic selections that can be referenced in plots and
layouts. Each selection gets a unique name (or the name you provide) and is
automatically added to the spec `params` during `mosaic_spec()`.
}
\section{Fields}{

\describe{
\item{\code{type}}{Selection type}

\item{\code{name}}{Selection parameter name}
}}

\examples{
\dontrun{
brush <- Selection$crossfilter()
single <- Selection$single()
}
}
\keyword{datasets}
