% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{store_mosaic_selection}
\alias{store_mosaic_selection}
\title{Persist a data.frame into an R environment}
\usage{
store_mosaic_selection(df, env = .GlobalEnv)
}
\arguments{
\item{df}{data.frame to store}

\item{env}{environment to store the selection in}
}
\value{
Name of the variable created
}
\description{
Persist a data.frame into an R environment
}
