% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{tick_format}
\alias{tick_format}
\title{Set tick format}
\usage{
tick_format(x = NULL, y = NULL)
}
\arguments{
\item{x}{X-axis tick format}

\item{y}{Y-axis tick format}
}
\description{
Set tick format
}
