% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_starter_template.R, R/ggmosaic.R,
%   R/ggmosaic_enhanced.R
\name{vconcat}
\alias{vconcat}
\title{Vertical concatenation of plots}
\usage{
vconcat(...)

vconcat(...)

vconcat(...)
}
\arguments{
\item{...}{ggmosaic plots or layout objects to concatenate vertically}
}
\value{
A layout specification

A layout object
}
\description{
Vertical concatenation of plots

Vertical concatenation

Vertical concatenation of plots
}
\examples{
\dontrun{
p1 <- ggmosaic(mtcars, aes(x = mpg)) + geom_histogram()
p2 <- ggmosaic(mtcars, aes(x = hp)) + geom_histogram()
vconcat(p1, p2) \%>\% render()
}
}
