% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_enhanced.R
\name{vspace}
\alias{vspace}
\title{Add vertical spacing}
\usage{
vspace(pixels = 10)
}
\arguments{
\item{pixels}{Number of pixels for vertical spacing}
}
\value{
A spacing element
}
\description{
Add vertical spacing
}
