% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_p1.R
\name{mos_fetch}
\alias{mos_fetch}
\title{Returns the mosaic database}
\usage{
mos_fetch(id_key)
}
\arguments{
\item{versionSet}{Version of the MOSAIC database}
}
\value{
mosaic database
}
\description{
This downloads the MOSAIC database as an S4 object
}
\examples{
mos_fetch("v1.0.0")
}
