% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{mosaicOutput}
\alias{mosaicOutput}
\title{Shiny output for Mosaic widget}
\usage{
mosaicOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable name}

\item{width, height}{CSS dimensions (e.g. '100\%', '400px') for the container.}
}
\value{
A Shiny UI output element for a Mosaic htmlwidget.
}
\description{
Use this in the UI to create a placeholder for the Mosaic visualization.
}
