% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderMosaic}
\alias{renderMosaic}
\title{Shiny render function for Mosaic}
\usage{
renderMosaic(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a call to \code{\link{mosaic}()}.}
}
\description{
Use this in the server to render the Mosaic visualization to the output.
}
