% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMosaic-package.R
\docType{package}
\name{rMosaic-package}
\alias{rMosaic}
\alias{rMosaic-package}
\title{rMosaic: Mosaic Shiny Integration}
\description{
Embed the Mosaic framework in R, allowing users to declaratively define visualizations (via JSON/YAML specs) and render interactive Mosaic plots in Shiny. Data queries are executed in DuckDB via R, enabling scalable linked visuals. Includes Mosaic JavaScript libraries v0.21.1.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TiRizvanov/rMosaic}
  \item Report bugs at \url{https://github.com/TiRizvanov/rMosaic/issues}
}

}
\author{
\strong{Maintainer}: Timur Rizvanov \email{timurr@bu.edu}

}
\keyword{internal}
