% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMosaic-package.R
\docType{package}
\name{rMosaic-package}
\alias{rMosaic}
\alias{rMosaic-package}
\title{rMosaic: Mosaic Shiny Integration}
\description{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TiRizvanov/rMosaic}
  \item Report bugs at \url{https://github.com/TiRizvanov/rMosaic/issues}
}

}
\author{
\strong{Maintainer}: Timur Rizvanov \email{timurr@bu.edu}

Authors:
\itemize{
  \item Edward C. Ruiz \email{ecr7407@gmail.com}
  \item Ruben Dries \email{rubendries@gmail.com} [reviewer]
}

}
\keyword{internal}
