% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSelection.R
\name{runMosaicExport}
\alias{runMosaicExport}
\title{Run a Mosaic app and export brush/query selections back to R}
\usage{
runMosaicExport(
  spec,
  specType = "auto",
  data,
  title = NULL,
  width = "100\%",
  height = "600px",
  selection_env = .GlobalEnv
)
}
\description{
Run a Mosaic app and export brush/query selections back to R
}
